# ARROW-1571: [C++][Compute] Optimize sorting integers in small value range