For integer array with limited value range, counting sort can be much
more efficient than comparison based algorithms. Detailed analysis in
Jira link https://issues.apache.org/jira/browse/ARROW-1571

Tested on x86 and Arm64 platforms, Arrow sorting kernel benchnmark
(values within -100 ~ +100) shows 6x to 15x performance boost.

NOTE: This approach scans for min/max values before choose appropriate
sorting algorithm. It's a waste of time if we finally find the value
range is too large to use counting sort. This time is trivial comparing
with sorting. But I see performance drop when there are many null values
in the array. Possibly due to unpredictable null value checking impacts
finding min/max values. In case of 50% null values, up to 6% performance
drop is observed in sorting big int64 arrays. Will follow up this issue.