@pitrou 2020-03-02
@cyb70289 Can you try this patch? It seems to fix the performance issue with `ArrayDataVisitor` for me (on gcc 7.4.0, Ubuntu 18.04, AMD Ryzen):
```c++
diff --git a/cpp/src/arrow/status.h b/cpp/src/arrow/status.h
index dbe1b7e6e..195ed71cc 100644
--- a/cpp/src/arrow/status.h
+++ b/cpp/src/arrow/status.h
@@ -436,6 +436,7 @@ namespace internal {
 // Extract Status from Status or Result<T>
 // Useful for the status check macros such as RETURN_NOT_OK.
 inline Status GenericToStatus(const Status& st) { return st; }
+inline Status GenericToStatus(Status&& st) { return std::move(st); }
 
 }  // namespace internal
 
```