@cyb70289 2020-03-02
> @cyb70289 Can you try this patch? It seems to fix the performance issue with `ArrayDataVisitor` for me (on gcc 7.4.0, Ubuntu 18.04, AMD Ryzen):

It works for me too. Cool!