@pitrou 2020-03-02
@cyb70289 Can you try to benchmark the commit I pushed?