@pitrou 2020-03-02
Note the trivial `GenericToStatus` change actually makes other benchmarks significantly faster.
For example Parquet decoding:
* before:
```
BM_ArrowBinaryPlain/DecodeArrow_Dense/65536            622135 ns       622032 ns         1124 bytes_per_second=604.752M/s
BM_ArrowBinaryPlain/DecodeArrowNonNull_Dense/65536     621900 ns       621815 ns         1126 bytes_per_second=604.962M/s
BM_ArrowBinaryPlain/DecodeArrow_Dict/65536            3356821 ns      3356261 ns          208 bytes_per_second=112.082M/s
BM_ArrowBinaryPlain/DecodeArrowNonNull_Dict/65536     3350232 ns      3349732 ns          209 bytes_per_second=112.3M/s
BM_ArrowBinaryDict/DecodeArrow_Dense/65536            2062856 ns      2062515 ns          339 bytes_per_second=182.386M/s
BM_ArrowBinaryDict/DecodeArrowNonNull_Dense/65536     2064331 ns      2063978 ns          339 bytes_per_second=182.257M/s
BM_ArrowBinaryDict/DecodeArrow_Dict/65536             3854847 ns      3854230 ns          182 bytes_per_second=97.6005M/s
BM_ArrowBinaryDict/DecodeArrowNonNull_Dict/65536      3844835 ns      3844182 ns          182 bytes_per_second=97.8556M/s
```
* after:
```
BM_ArrowBinaryPlain/DecodeArrow_Dense/65536            633608 ns       633510 ns         1106 bytes_per_second=593.795M/s
BM_ArrowBinaryPlain/DecodeArrowNonNull_Dense/65536     635779 ns       635673 ns         1099 bytes_per_second=591.775M/s
BM_ArrowBinaryPlain/DecodeArrow_Dict/65536            2169264 ns      2168904 ns          324 bytes_per_second=173.44M/s
BM_ArrowBinaryPlain/DecodeArrowNonNull_Dict/65536     2178579 ns      2178265 ns          320 bytes_per_second=172.695M/s
BM_ArrowBinaryDict/DecodeArrow_Dense/65536            1217441 ns      1217257 ns          575 bytes_per_second=309.035M/s
BM_ArrowBinaryDict/DecodeArrowNonNull_Dense/65536     1217940 ns      1217727 ns          577 bytes_per_second=308.916M/s
BM_ArrowBinaryDict/DecodeArrow_Dict/65536             2569992 ns      2569625 ns          273 bytes_per_second=146.393M/s
BM_ArrowBinaryDict/DecodeArrowNonNull_Dict/65536      2698939 ns      2698444 ns          255 bytes_per_second=139.404M/s
```