@pitrou 2020-03-02
(R CI failure is unrelated)