@wesm 2020-03-02
Pretty amazing the GenericToStatus performance improvement. Gives me more of a feeling of urgency to track and visualize all of our benchmarks continuously