@cyb70289 2020-03-03
Thanks @pitrou @wesm , amazing improvement with a trivial change!
I guess that's why there will be a whole book for C++11 move semantics :)
http://www.cppmove.com