@wesm 2020-03-03
@pitrou this patch broke Appveyor it seems https://ci.appveyor.com/project/ApacheSoftwareFoundation/arrow/builds/31195004/job/y7mfa2xtv8yxy0p0#L1052. 

https://issues.apache.org/jira/browse/ARROW-7992